+++ src/Generics/Deriving/Monoid.hs
@@ -12,6 +12,16 @@ module Generics.Deriving.Monoid (module Generics.Deriving.Monoid.Internal) where
 import Generics.Deriving.Monoid.Internal
 import Generics.Deriving.Semigroup (GSemigroup(..))
 
+#if MIN_VERSION_base(4,9,0)
+import Data.Semigroup (WrappedMonoid)
+#endif
+
 instance GSemigroup a => GMonoid (Maybe a) where
   gmempty = Nothing
   gmappend = gsappend
+
+#if MIN_VERSION_base(4,9,0)
+instance GMonoid m => GMonoid (WrappedMonoid m) where
+  gmempty  = gmemptydefault
+  gmappend = gmappenddefault
+#endif